feat: migrate to Zephyr 4.2.0
* replace deprecated `SB_CONFIG_MCUBOOT_MODE_SWAP_WITHOUT_SCRATCH` with `SB_CONFIG_MCUBOOT_MODE_SWAP_USING_MOVE`
From
https://docs.zephyrproject.org/latest/releases/migration-guide-4.1.html:
```
The Kconfig ``SB_CONFIG_MCUBOOT_MODE_SWAP_WITHOUT_SCRATCH`` has been deprecated and replaced
with ``SB_CONFIG_MCUBOOT_MODE_SWAP_USING_MOVE``, applications should be updated to select this
new symbol if they were selecting the old symbol.
```

* set `mgmt_event` type to `uint64_t`
From https://docs.zephyrproject.org/latest/releases/migration-guide-4.2.html:
The management event type is changed from `uint32_t` to `uint64_t`

* explicitly enable `POSIX_C_LANG_SUPPORT_R`
The General ISO C Library is needed to get the ISO 8601 timestamps we
need for the deployment logs.

* use cmsis-6
https://github.com/zephyrproject-rtos/example-application/commit/06bbb0a5e5adc5b6da0c0bb16919cebea931ee7e

* increase log buffer
The logging suddnely stop during a deployment. Inrease from default 1028
to 2046

* disable spin validation
From the Kconfig:
  There's a spinlock validation framework available when asserts are
  enabled. It adds a relatively hefty overhead (about 3k or so) to
  kernel code size, don't use on platforms known to be small.
If we have this enabled we will get an error when attempting to call the
reboot callback saying `... Context switching while holding lock!`

Ticket: MEN-8638
Changelog: Ticket

+++ src/utils/netup.c
@@ -74,7 +74,7 @@ wifi_connect(struct net_if *iface) {
 #endif
 
 static void
-event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event, struct net_if *iface) {
+event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event, struct net_if *iface) {
     int i = 0;
 
     if (mgmt_event != NET_EVENT_IPV4_ADDR_ADD) {
